+++ src/const.py
@@ -1,0 +1,17 @@
+"""
+Configuration handling of the integration driver.
+
+:copyright: (c) 2023-2026 by Unfolded Circle ApS.
+:license: Mozilla Public License Version 2.0, see LICENSE for more details.
+"""
+
+from enum import Enum
+from typing import Any, Type
+
+BROWINS_APP_ID = "albaintor.UC-Remote.tv"
+
+
+def filter_attributes(attributes, attribute_type: Type[Enum]) -> dict[str, Any]:
+    """Filter attributes based on an Enum class."""
+    valid_keys = {e.value for e in attribute_type}
+    return {k: v for k, v in attributes.items() if k in valid_keys}